move combination.hpp and update PFEG
