+++ RecoParticleFlow/PFProducer/src/combination.hpp
